Update project status in README
The ambiguity of the status of this project has been touched upon in
many recent issues. This update removes ambiguity about this service and
this software. The outcome here is that phant will officially be
de-prioritized; however, this signals to the community surrounding phant
that new maintainership would be welcome.

Fixes: #222, #221, #219, #208, #188, #187, #179